JWT refresh endpoint
Implemented the refresh token. This endpoint takes in a refresh token
and an audience. The backend checks that the refresh token was created
and assigned to the audience given in the POST request. If this matches
and the token is valid, a new access token is issued to the user.