Added server-side hooks
Added svelte server-side hooks. This is a svelte feature which allows
every svelte server-side rendered request to pass through a hook. This
can then manipulate every request.

This will be used to check if the response is 401 (Unauthorised), in
which case it is very likely that the users accessToken has expired. If
so attempt to refresh their token then retry initial request.